backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.24.0
sqlalchemy>=2.0.0
pydantic>=2.11.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pyodbc>=5.0.0
python-dotenv>=1.0.0
alembic>=1.12.0
pymssql==2.2.8
pytest==7.4.3
httpx==0.25.2
ldap3>=2.9.1
slowapi>=0.1.9
redis>=5.0.1
cryptography>=41.0.7 